[PF-34] - Simulation
- extract common functions to abstract Simulation class
- fix Event id after merge
