+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@netgrif/petriflow",
-    "version": "1.3.5",
+    "version": "1.3.6",
     "description": "Javascript / Typescript library of Petriflow objects",
     "main": "dist/petriflow.js",
     "module": "dist/petriflow.esm.js",
