+++ src/lib/simulation/index.ts
@@ -1,1 +1,1 @@
-export * from './simulation';
+export * from './basic-simulation';
